Move Vue components to frappejs
