+++ src/App.vue
@@ -1,45 +1,26 @@
 <template>
   <div id="app">
-    <frappe-desk v-if="showDesk">
+    <frappe-desk v-if="showDesk" :sidebarConfig="sidebarConfig">
       <router-view />
     </frappe-desk>
     <router-view v-else name="setup" />
-    <frappe-modal ref="modal" :show="showModal" v-bind="modalOptions" @close-modal="showModal = false"/>
+    <frappe-modal ref="modal" :show="modalVisible" v-bind="modalOptions" @close-modal="modalVisible = false"/>
   </div>
 </template>
 
 <script>
 import Vue from 'vue';
 import Observable from 'frappejs/utils/observable';
-import Desk from '@/components/Desk';
-import Modal from '@/components/Modal';
-
-const Bus = new Observable();
-
-Vue.use({
-  // enable use of this.$modal in every component
-  // this also keeps only one modal in the DOM at any time
-  // which is the recommended way by bootstrap
-  install (Vue) {
-    Vue.prototype.$modal = {
-      show(options) {
-        Bus.trigger('showModal', options);
-      },
-
-      hide() {
-        Bus.trigger('hideModal');
-      }
-    }
-  }
-});
+import Desk from 'frappejs/ui/components/Desk';
+import Modal from 'frappejs/ui/components/Modal';
+import sidebarConfig from './sidebarConfig';
 
 export default {
   name: 'App',
   data() {
     return {
       showDesk: true,
-      showModal: false,
-      modalOptions: {}
+      sidebarConfig
     }
   },
   components: {
@@ -53,18 +34,6 @@ export default {
     } else {
       this.showDesk = true;
     }
-  },
-  mounted() {
-    Bus.on('showModal', (options = {}) => {
-      this.showModal = true;
-      this.modalOptions = options;
-    });
-
-    Bus.on('hideModal', () => {
-      this.showModal = false;
-    });
-
-    window.frappeModal = this.$modal;
   }
 }
 </script>
