--- src/components/Main.vue
@@ -1,5 +1,0 @@
-<template>
-    <main class="frappe-main col-10">
-        <slot></slot>
-    </main>
-</template>
