--- src/components/Navbar.vue
@@ -1,7 +1,0 @@
-<template>
-    <nav class="frappe-navbar navbar navbar-light bg-light row no-gutters border-bottom border-top">
-        <form class="form-inline col-4 pr-3">
-            <input type="search" name="search" class="form-control shadow-none w-100" placeholder="Search...">
-        </form>
-    </nav>
-</template>
