--- src/components/NotFound.vue
@@ -1,11 +1,0 @@
-<template>
-    <div class="frappe-not-found d-flex flex-column align-items-center justify-content-center">
-        <h1>404</h1>
-        <p>Not Found</p>
-    </div>
-</template>
-<style>
-.frappe-not-found {
-    height: 100%;
-}
-</style>
