--- src/components/controls/Code.vue
@@ -1,11 +1,0 @@
-<script>
-import Text from './Text';
-export default {
-    extends: Text,
-    computed: {
-        inputClass() {
-            return ['text-monospace'];
-        }
-    }
-}
-</script>
