--- src/components/controls/Currency.vue
@@ -1,6 +1,0 @@
-<script>
-import Float from './Float';
-export default {
-    extends: Float
-}
-</script>
