--- src/components/controls/Data.vue
@@ -1,8 +1,0 @@
-<script>
-import Base from './Base';
-
-export default {
-    name: 'Data',
-    extends: Base
-}
-</script>
