--- src/components/controls/DynamicLink.vue
@@ -1,14 +1,0 @@
-<script>
-import frappe from 'frappejs';
-import Link from './Link';
-
-export default {
-    extends: Link,
-    inject: ['dynamicLinkTarget'],
-    methods: {
-        getTarget() {
-            return this.dynamicLinkTarget(this.docfield.references);
-        }
-    }
-}
-</script>
