--- src/components/controls/Int.vue
@@ -1,13 +1,0 @@
-<script>
-import Float from './Float';
-
-export default {
-    extends: Float,
-    methods: {
-        parse(value) {
-            const parsedValue = parseInt(value);
-            return isNaN(parsedValue) ? 0 : parsedValue;
-        }
-    }
-}
-</script>
