--- src/components/controls/Password.vue
@@ -1,18 +1,0 @@
-<script>
-import Base from "./Base";
-
-export default {
-  name: "Password",
-  extends: Base,
-  methods: {
-    getInputAttrs() {
-      return {
-        id: this.id,
-        type: 'password',
-        value: this.value,
-        required: this.docfield.required
-      };
-    },
-  }
-};
-</script>
