--- src/components/controls/Select.vue
@@ -1,30 +1,0 @@
-<script>
-import Base from './Base';
-export default {
-    extends: Base,
-    methods: {
-        getInputTag() {
-            return 'select';
-        },
-        getInputAttrs() {
-            return {
-                id: this.id,
-                required: this.docfield.required
-            };
-        },
-        getInputChildren(h) {
-            return this.docfield.options.map(option =>
-                h('option', {
-                    attrs: {
-                        key: option,
-                        selected: option === this.value
-                    },
-                    domProps: {
-                        textContent: option
-                    }
-                })
-            );
-        }
-    }
-}
-</script>
