--- src/components/controls/Text.vue
@@ -1,23 +1,0 @@
-<script>
-import Base from './Base';
-export default {
-    extends: Base,
-    methods: {
-        getInputTag() {
-            return 'textarea';
-        },
-        getInputAttrs() {
-            return {
-                id: this.id,
-                required: this.docfield.required,
-                rows: 3
-            };
-        },
-        getDomProps() {
-            return {
-                value: this.value
-            }
-        }
-    }
-}
-</script>
