+++ src/router/index.js
@@ -1,52 +1,33 @@
-import Vue from 'vue'
-import Router from 'vue-router'
-import ListAndForm from '@/pages/ListAndForm'
-import SetupWizard from '@/pages/SetupWizard/SetupWizard'
-import Report from '@/components/Report'
+import Vue from 'vue';
+import Router from 'vue-router';
+import coreRoutes from 'frappejs/ui/routes';
 
-Vue.use(Router)
+import SetupWizard from '../pages/SetupWizard';
+import Report from 'frappejs/ui/pages/Report';
+import reportViewConfig from '../../reports/view';
 
-export default new Router({
-  routes: [
-    {
-      path: '/list/:doctype',
-      name: 'List',
-      component: ListAndForm,
-      props: true
-    },
-    {
-      path: '/edit/:doctype/:name',
-      name: 'Form',
-      component: ListAndForm,
-      props: true
-    },
-    {
-      path: '/setup-wizard',
-      name: 'SetupWizard',
-      components: {
-        setup: SetupWizard
-      }
-    },
-    {
-      path: '/report/:reportName',
-      name: 'Report',
-      component: Report,
-      props: true
+Vue.use(Router);
+
+const routes = [].concat(coreRoutes, [
+  {
+    path: '/setup-wizard',
+    name: 'SetupWizard',
+    components: {
+      setup: SetupWizard
     }
-    // {
-    //   path: '/table/:doctype',
-    //   name: '',
-    //   component: ''
-    // },
-    // {
-    //   path: '/print/:doctype/:name',
-    //   name: '',
-    //   component: ''
-    // },
-    // {
-    //   path: '/new/:doctype',
-    //   name: '',
-    //   component: ''
-    // }
-  ]
-})
+  },
+  {
+    path: '/report/:reportName',
+    name: 'Report',
+    component: Report,
+    props: (route) => {
+      const { reportName } = route.params;
+      return {
+        reportName,
+        reportConfig: reportViewConfig[reportName] || null
+      };
+    }
+  }
+]);
+
+export default new Router({ routes });
