+++ src/sidebarConfig.js
@@ -39,4 +39,4 @@ export default [
       }
     ]
   }
-]
+];
